chore(frontend): Migrate from Remix to React Router 7 (#5304)
